+++ src/model/Dict.ts
@@ -1,0 +1,15 @@
+/**
+ * 单词本
+ */
+
+export default class Dict {
+  http;
+
+  constructor(http){
+    this.http = http;
+  }
+
+  add(word) {
+    return this.http.addWord(word);
+  }
+}
