+++ src/model/interface.ts
@@ -1,0 +1,16 @@
+/**
+ * 单词本
+ */
+export interface Dict{
+  /**
+   * 添加单词
+   */
+  add:(word: string)=> void;
+}
+
+/**
+ * 用户设置
+ */
+export interface Setting {
+
+}
